+++ I wanna Siv3D/OpenSiv3D_0.6.161/StageManager/StageManager.cpp
@@ -145,11 +145,11 @@ namespace Iwanna {
 	}
 
 	Vec2 StageManager::parsePos(const JSON& json) {
-		return Vec2{ json[0].get<int32>() / oneTileSize, json[1].get<int32>() / oneTileSize };
+		return Vec2{ json[0].get<double>() / oneTileSize, json[1].get<double>() / oneTileSize };
 	}
 
 	Vec2 StageManager::parseIntactPos(const JSON& json) {
-		return Vec2{ json[0].get<int32>(), json[1].get<int32>()};
+		return Vec2{ json[0].get<double>(), json[1].get<double>() };
 	}
 
 	void StageManager::update() {
